--- clinicadl/utils/cluster/api/__init__.py
@@ -1,13 +1,0 @@
-from .auto_master_addr_port import AutoMasterAddressPort
-from .base import API
-from .default import DefaultAPI
-from .slurm import SlurmAPI
-from .torchelastic import TorchElasticAPI
-
-__all__ = [
-    "API",
-    "AutoMasterAddressPort",
-    "DefaultAPI",
-    "SlurmAPI",
-    "TorchElasticAPI",
-]
