+++ clinicadl/utils/dictionary/suffixes.py
@@ -3,8 +3,3 @@ TXT = ".txt"
 JSON = ".json"
 LOG = ".log"
 PT = ".pt"
-PTH = ".pth"
-TAR = ".tar"
-NII = ".nii"
-GZ = ".gz"
-NII_GZ = NII + GZ
