+++ clinicadl/utils/dictionary/words.py
@@ -9,11 +9,9 @@ BEST = "best"
 BIDS = "bids"
 CALLBACKS = "callbacks"
 CAPS = "caps"
-CAPS_DATASET = "caps_dataset"
 CHECKPOINT = "checkpoint"
 CHECKPOINTS = CHECKPOINT + "s"
 CLINICADL = "clinicadl"
-COLUMNS = "columns"
 COMPUTATIONAL = "computational"
 CONFIG = "config"
 COPY = "copy"
@@ -26,35 +24,27 @@ DATALOADER = "dataloader"
 DATASET = "dataset"
 DATASET_ID = DATASET + "_id"
 DEBUG = "debug"
-DESCRIPTION = "description"
 DEFAULT = "default"
 DETAILS = "details"
-DF = "df"
 DIRECTION = "direction"
 EXEC = "exec"
 ENVIRONMENT = "environment"
 ERROR = "error"
 EPOCH = "epoch"
 EXCLUDE = "exclude"
-FIRST_INDEX = "first_idx"
 FINAL = "final"
 FILE_TYPE = "file_type"
 FOLD = "fold"
 GROUP = "group"
-GROUPS = GROUP + "s"
 GPU = "gpu"
 ID = "id"
 IMAGE = "image"
 IMAGE_PATH = IMAGE + "_path"
 INCLUDE = "include"
-INFORMATION = "information"
 INFO = "info"
-KFOLD = "k" + FOLD
 LABEL = "label"
-LAST_INDEX = "last_idx"
 LOGS = "logs"
 LOSS = "loss"
-LOSS_METRIC = "LossMetric"
 MASK = "mask"
 MAPS = "maps"
 MEAN = "mean"
@@ -66,25 +56,17 @@ NN = "nn"
 N_SAMPLES = "n_samples"
 NAME = "name"
 NAME_ = "name_"
-OBJECT = "object"
-OTHER = "other"
 OPTIMIZATION = "optimization"
 OPTIMIZER = "optimizer"
 OUTPUT = "output"
 PARTICIPANT = "participant"
-PARTICIPANTS_SESSIONS = "participantsXsessions"
 PARTICIPANT_ID = PARTICIPANT + "_" + ID
-PERFORMANCE = "performance"
 POSITION = "position"
-PREPARATION = "preparation"
 PREDICTION = "prediction"
-PREDICTIONS = PREDICTION + "s"
-PREPROCESSING = "preprocessing"
 PROPORTION = "proportion"
 READER = "reader"
 RESULTS = "results"
 RUN = "run"
-RUNS = RUN + "s"
 SAMPLE = "sample"
 SAMPLE_POSITION = SAMPLE + "_" + POSITION
 TYPE = "type"
@@ -93,7 +75,6 @@ SCALER = "scaler"
 SEX = "sex"
 SESSION = "session"
 SESSION_ID = SESSION + "_" + ID
-SELECTION_METRICS = "selection_metrics"
 SINGLE = "single"
 SLICE = "slice"
 SLICE_DIRECTION = SLICE + "_" + DIRECTION
@@ -102,23 +83,15 @@ SQUEEZE = "squeeze"
 STATISTIC = "statistic"
 STATE = "state"
 STD = "std"
-STOP = "stop"
-SUBJECTS = "subjects"
 SUMMARY = "summary"
 TEST = "test"
 TIME = "time"
 TMP = "tmp"
-TORCH = "torch"
 TRAIN = "train"
 TRAINING = "training"
 TRANSFORMATION = "transformation"
-TRANSFORMS = "transforms"
-TENSORBOARD = "tensorboard"
 TENSOR = "tensor"
-TENSORS = TENSOR + "s"
-TENSOR_CONVERSION = "tensor_conversion"
 LEARNING_RATES = "learning_rates"
-TYPE = "type"
 VALIDATION = "validation"
 VALUE = "value"
 WARNING = "warning"
